refactors label annotation tools
